+++ internal/cmd/db_remove.go
@@ -25,7 +25,7 @@ var dbRemoveCommand = &cobra.Command{
 
 		// get all of the docker containers by name
 		script := scripts.New(mp, machine)
-		output, err := script.Run(scripts.DockerListContainerNames)
+		output, err := script.Run(false, scripts.DockerListContainerNames)
 		if err != nil {
 			return err
 		}
@@ -52,7 +52,7 @@ var dbRemoveCommand = &cobra.Command{
 		}
 
 		if remove {
-			_, err = script.Run(fmt.Sprintf(scripts.FmtDockerRemoveContainer, container))
+			_, err = script.Run(false, fmt.Sprintf(scripts.FmtDockerRemoveContainer, container))
 			if err != nil {
 				return err
 			}